Include more things in the content hash
I'm actually thinking it might be better to default to including
everything in the hash and perhaps excluding a few things which
shouldn't count. My goal had been to limit this to the things that
definitely would change the resulting image without including
things which should reasonably be expected not to do so (such as
secrets), though I'm no longer sure if that's a sensible aim.